Moving unit testing into SearchParameters test class
